+++ vscode-extensiones/ext1/summary1/CHANGELOG.md
@@ -8,3 +8,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Initial release
 - Ajustes en la forma como toma y genera el resumen.
+- Se agrega directorios y archivos excluidos por defecto